# api/src/swagger.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Rondeau API",
        "description": "Primary interface for Rondeau app functions.",
        "version": 0.1
    },
    "servers": [
        {
            "url": "https://rondeau.dillonschwertz.dev/v0",
            "description": "Main"
        },
        {
            "url": "http://localhost:5173/v0",
            "description": "Local"
        }
    ],
    "components": {
        "securitySchemes": {
            "default-authorizer": {
                "type": "apiKey",
                "in": "header",
                "name": "Cookie",
                "x-amazon-apigateway-authtype": "custom",
                "x-amazon-apigateway-authorizer": {
                    "type": "token",
                    "authorizerUri": "arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:446725842066:function:ApiAuthorizer/invocations",
                    "enableSimpleResponses": true,
                    "authorizerResultTtlInSeconds": 60,
                    "identitySource": "method.request.header.Cookie"
                }
            }
        }
    },
    "paths": {
        "/auth/status": {
            "get": {
                "summary": "Get user information from id token",
                "description": "This resource triggers the api-auth-status Lambda which decodes the id token from the cookie and returns an object defining important user attributes\n",
                "security": [],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "type": "aws_proxy",
                    "httpMethod": "POST",
                    "uri": {
                        "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:ApiAuthStatus/invocations"
                    }
                }
            }
        },
        "/auth/callback": {
            "get": {
                "summary": "Callback function for Cognito authentication.",
                "description": "This resource triggers the api-auth-callback Lambda which exchanges the code sent by Cognito for id, access, and refresh tokens which are stored in cookies.\n",
                "security": [],
                "responses": {
                    "302": {
                        "description": "Successful token exchange",
                        "headers": {
                            "location": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing parameters for token exchange",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed token exchange",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "x-amazon-apigateway-auth": {
                    "type": "none"
                },
                "x-amazon-apigateway-integration": {
                    "type": "aws_proxy",
                    "httpMethod": "POST",
                    "uri": {
                        "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:ApiAuthCallback/invocations"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "summary": "Logout function for Cognito authentication.",
                "description": "This resource triggers the api-auth-logout Lambda which invalidates existing tokens and clears the id, access, and refresh tokens from browser cookies.\n",
                "security": [],
                "responses": {
                    "302": {
                        "description": "Successful logout",
                        "headers": {
                            "location": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "x-amazon-apigateway-auth": {
                    "type": "none"
                },
                "x-amazon-apigateway-integration": {
                    "type": "aws_proxy",
                    "httpMethod": "POST",
                    "uri": {
                        "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:ApiAuthLogout/invocations"
                    }
                }
            }
        },
        "/status": {
            "get": {
                "summary": "Get a simple status response for the API",
                "description": "This resource triggers the api-status Lambda which simply returns a 200 response indicated a functional API.\n",
                "security": [
                    {
                        "default-authorizer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Functional API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "x-amazon-apigateway-integration": {
                    "type": "aws_proxy",
                    "httpMethod": "POST",
                    "uri": {
                        "Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:ApiStatus/invocations"
                    }
                }
            }
        }
    }
}